refactor: extract team default tools into dedicated _default_tools.py module
Mirror the Agent pattern by splitting team/_tools.py into infrastructure
(_tools.py) and default tool factories (_default_tools.py). This resolves
the asymmetry where Agent had 11 trait modules and Team only had 9.
